% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{effective_traits}
\alias{effective_traits}
\title{Effective numbers of traits}
\usage{
effective_traits(intercept, rg)
}
\arguments{
\item{intercept}{q-by-q intercept matrix \eqn{\Gamma} (scaled to unit
diagonal internally).}

\item{rg}{q-by-q genetic correlation matrix (or a genetic covariance,
converted to correlation).}
}
\description{
\eqn{q_P = q^2/\sum\Gamma_{ab}^2} (sample overlap times phenotypic
correlation), \eqn{q_G = q^2/\sum r_{g,ab}^2} and
\eqn{q_{PG} = q^2/\sum \Gamma_{ab} r_{g,ab}}. Null standard errors depend
only on \eqn{q_P}; power at large enrichment is governed by \eqn{q_G}
(theory supplement S9.1). A single \eqn{q_{eff}} is only a null summary.
}
