% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panukb.R
\name{http_get}
\alias{http_get}
\title{HTTP GET (optionally a byte range) with retries via the curl binary}
\usage{
http_get(url, dest, range = NULL, tries = 4)
}
\description{
HTTP GET (optionally a byte range) with retries via the curl binary
}
\keyword{internal}
