% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ld_apply}
\alias{ld_apply}
\title{Apply an LD matrix given as a matrix or a list of diagonal blocks}
\usage{
ld_apply(ld, X, fun = function(R, X) R \%*\% X)
}
\description{
Apply an LD matrix given as a matrix or a list of diagonal blocks
}
\keyword{internal}
