% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ld_bdiag}
\alias{ld_bdiag}
\title{Bind a list of LD blocks into one (sparse) block-diagonal matrix}
\usage{
ld_bdiag(blocks)
}
\arguments{
\item{blocks}{List of LD blocks.}
}
\description{
Bind a list of LD blocks into one (sparse) block-diagonal matrix
}
