% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_from_blocks}
\alias{ld_from_blocks}
\title{LD reference from a list of independent LD blocks}
\usage{
ld_from_blocks(blocks, snp = NULL, N = Inf)
}
\arguments{
\item{blocks}{List of LD block matrices (zero LD between blocks).}

\item{snp}{Optional SNP metadata data frame.}

\item{N}{Reference panel sample size (for r-squared bias correction;
\code{Inf} for population LD).}
}
\description{
LD reference from a list of independent LD blocks
}
