% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_from_genotypes}
\alias{ld_from_genotypes}
\title{LD reference computed on the fly from a genotype matrix}
\usage{
ld_from_genotypes(G, block_id, snp = NULL)
}
\arguments{
\item{G}{Individuals-by-SNPs genotype dosage matrix (0/1/2), SNPs in
genomic order.}

\item{block_id}{Block label per SNP (see [assign_blocks()]).}

\item{snp}{Optional SNP metadata data frame.}
}
\description{
LD reference computed on the fly from a genotype matrix
}
