% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_ld_ar1}
\alias{make_ld_ar1}
\title{Toy LD: block-diagonal AR(1) correlation}
\usage{
make_ld_ar1(block_sizes, rho = 0.5)
}
\arguments{
\item{block_sizes}{Integer vector of block sizes.}

\item{rho}{AR(1) correlation between adjacent SNPs within a block.}
}
\value{
List of LD blocks.
}
\description{
Toy LD: block-diagonal AR(1) correlation
}
