% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcp.R
\name{MASS_ginv}
\alias{MASS_ginv}
\title{Moore-Penrose inverse (fallback for singular Hessians)}
\usage{
MASS_ginv(X, tol = sqrt(.Machine$double.eps))
}
\description{
Moore-Penrose inverse (fallback for singular Hessians)
}
\keyword{internal}
