% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panukb.R
\name{panukb_manifest}
\alias{panukb_manifest}
\title{Read the Pan-UKB phenotype manifest}
\usage{
panukb_manifest(
  path =
    "https://pan-ukb-us-east-1.s3.amazonaws.com/sumstats_release/phenotype_manifest.tsv.bgz"
)
}
\arguments{
\item{path}{Local path or URL of \code{phenotype_manifest.tsv.bgz}
(default: the public S3 copy).}
}
\value{
Data frame.
}
\description{
Read the Pan-UKB phenotype manifest
}
