% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{quadform_var}
\alias{quadform_var}
\title{Exact variance of a weighted quadratic form in Z-scores}
\usage{
quadform_var(A, u, R, G, Gamma, Tm)
}
\arguments{
\item{A}{Symmetric m-by-m matrix of the quadratic form.}

\item{u}{Length-q trait weights.}

\item{R}{LD matrix.}

\item{G}{\eqn{R S_w R} (genetic part of the Z covariance, per unit of
\eqn{t_{ab}}).}

\item{Gamma}{q-by-q intercept matrix.}

\item{Tm}{q-by-q scale matrix \eqn{t_{ab} = \sqrt{n_a n_b} h_{ab}/M}.}
}
\description{
Theorem 3.10 / eq. (7.1) of the theory supplement: for Gaussian Z with
\eqn{Cov(vec Z) = \Gamma \otimes R + T \otimes G},
\deqn{Var(\sum_a u_a z_a' A z_a) = 2[\tau_{RR} u'(\Gamma\circ\Gamma)u +
  2\tau_{RG} u'(\Gamma\circ T)u + \tau_{GG} u'(T\circ T)u],}
\eqn{\tau_{XY} = tr(AXAY)}.
}
