% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panukb.R
\name{read_tbi}
\alias{read_tbi}
\title{Read a tabix (.tbi) index}
\usage{
read_tbi(path)
}
\arguments{
\item{path}{Local path or URL of the \code{.tbi}.}
}
\value{
List with sequence names and, per sequence, the bin chunks and
  linear index (virtual file offsets as doubles).
}
\description{
Read a tabix (.tbi) index
}
