% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panukb.R
\name{reg2bins}
\alias{reg2bins}
\title{UCSC binning scheme bins overlapping [beg, end) (0-based)}
\usage{
reg2bins(beg, end)
}
\description{
UCSC binning scheme bins overlapping [beg, end) (0-based)
}
\keyword{internal}
