% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcp.R
\name{screen_snps}
\alias{screen_snps}
\title{Screen SNPs for LRCP on estimated enrichment}
\usage{
screen_snps(w, threshold = 0.5, z = NULL, z_min = -Inf)
}
\arguments{
\item{w}{Enrichment estimates.}

\item{threshold}{Cutoff \eqn{t}.}

\item{z}{Optional enrichment z-scores; if given, SNPs must also have
\code{z >= z_min}.}

\item{z_min}{Minimum z-score.}
}
\description{
Chapter 6 prescreening: \eqn{\rho_{ij}} is set to 0 and not estimated if
\eqn{w_i < t} or \eqn{w_j < t}.
}
