% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{se_lrcp}
\alias{se_lrcp}
\title{Standard error of a distal pair covariance estimate (eq. 9.1)}
\usage{
se_lrcp(wi, wj = wi, C = 0, n, h2, M = 1200000, qe)
}
\arguments{
\item{wi, wj}{Enrichment of the two clumps.}

\item{C}{Covariance \eqn{C_{ij} = \rho_{ij}\sqrt{w_i w_j}}.}

\item{n, h2}{Typical GWAS sample size and heritability.}

\item{M}{Number of SNPs fixing the per-SNP scale (\eqn{s = n h^2/M}).}

\item{qe}{Effective trait numbers from [effective_traits()] (named
vector), or one number used for all three.}
}
\description{
Standard error of a distal pair covariance estimate (eq. 9.1)
}
