% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{q_eff}
\alias{q_eff}
\title{Effective number of traits for a SNP pair}
\usage{
q_eff(Mk, Ml = Mk)
}
\arguments{
\item{Mk, Ml}{Trait covariance matrices from [expected_trait_cov()].}
}
\description{
A single-number summary that is exact for null standard errors only; for
power use [effective_traits()] (\eqn{q_P}, \eqn{q_G}, \eqn{q_{PG}}).
}
